Refactors analysis service & checklist schema
