+++ packages/shared/lib/utils/openai.ts
@@ -112,7 +112,9 @@ Format your response as a JSON object with the following structure:
   },
   "fileChecklists": [
     {
+      "id": "file-1",
       "filename": "path/to/file.ts",
+      "explanation": "Update the checklist for this file generated by OpenAI",
       "checklistItems": [
         {
           "id": "item-1",
